refactor: move table_info_checker to unithelper package
